refactor(playground): playground/shared パッケージで3環境の共通コードを統一
3環境（Vue3/Nuxt3/Nuxt4）で複製されていたページ本文・CSS・vee-validate
初期化を `playground/shared` ワークスペースパッケージに集約した。
各環境の pages/ は薄いラッパーに削減され、修正が1ファイルで完結する。
