+++ playground/nuxt3/nuxt.config.ts
@@ -23,7 +23,10 @@ export default defineNuxtConfig({
     minazukiUi: {
         theme: 'light'
     },
-    css: ['~/assets/css/playground.css'],
+    css: ['playground-shared/styles/playground.css'],
+    build: {
+        transpile: ['playground-shared']
+    },
     devServer: {
         port: 5175
     },
