--- playground/nuxt4/app/assets/css/playground.css
@@ -1,71 +1,0 @@
-body {
-    font-family: sans-serif;
-    color: var(--color-theme-text-primary);
-    background-color: var(--color-theme-bg-primary);
-    transition: background-color 0.2s, color 0.2s;
-}
-
-.pg-layout {
-    display: grid;
-    grid-template-rows: auto 1fr;
-    min-height: 100dvh;
-}
-
-.pg-header {
-    display: flex;
-    gap: 24px;
-    align-items: center;
-    padding: 12px 24px;
-    background-color: var(--color-theme-bg-secondary);
-    border-bottom: 1px solid var(--color-theme-border);
-}
-
-.pg-header-title {
-    margin-right: auto;
-    font-size: 1.1rem;
-    font-weight: bold;
-}
-
-.pg-nav {
-    display: flex;
-    gap: 8px;
-}
-
-.pg-nav a {
-    padding: 4px 12px;
-    font-size: 0.9rem;
-    color: var(--color-theme-text-primary);
-    text-decoration: none;
-    border-radius: 4px;
-}
-
-.pg-nav a:hover,
-.pg-nav a.router-link-active {
-    background-color: var(--color-theme-bg-select);
-}
-
-.pg-main {
-    width: 100%;
-    max-width: 960px;
-    padding: 24px;
-    margin: 0 auto;
-}
-
-.pg-section {
-    margin-bottom: 40px;
-}
-
-.pg-section h2 {
-    padding-bottom: 8px;
-    margin-bottom: 16px;
-    font-size: 1rem;
-    font-weight: bold;
-    border-bottom: 1px solid var(--color-theme-border);
-}
-
-.pg-row {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 12px;
-    align-items: flex-start;
-}
