+++ playground/nuxt4/app/plugins/init-validate.ts
@@ -1,10 +1,5 @@
-import { defineRule } from 'vee-validate';
-import { all } from '@vee-validate/rules';
-import { initValidate } from 'minazuki-ui';
+import { setupValidate } from 'playground-shared/validate';
 
 export default defineNuxtPlugin(() => {
-    Object.entries(all).forEach(([name, rule]) => {
-        defineRule(name, rule);
-    });
-    initValidate();
+    setupValidate();
 });
