+++ playground/nuxt4/package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "minazuki-ui": "workspace:*",
+    "playground-shared": "workspace:*",
     "@acab/reset.css": "^0.11",
     "@pinia/nuxt": "^0.10",
     "@vee-validate/rules": "^4",
