+++ playground/shared/pages/HomePage.vue
