+++ playground/vue/package.json
@@ -11,6 +11,7 @@
     "type-check": "vue-tsc --build --force"
   },
   "dependencies": {
+    "playground-shared": "workspace:*",
     "@acab/reset.css": "^0.11",
     "@unhead/vue": "^1",
     "@vee-validate/rules": "^4",
